+++ src/util.js
@@ -4,11 +4,16 @@ import path from 'path';
 
 import { glob } from 'glob';
 
+const CODE_REG = /^```\w*$(.+)```/ms;
 const JSON_REG = /([{[].+[}\]])/s;
 
 export async function loadTemplates(dir) {
   const result = {};
   const files = await glob(path.join(dir, '*.md'));
+
+  if (!files.length) {
+    throw new Error(`No templates found in: ${dir}.`);
+  }
 
   for (let file of files) {
     const base = path.basename(file, '.md');
@@ -16,15 +21,6 @@ export async function loadTemplates(dir) {
   }
 
   return result;
-}
-
-export function parse(content) {
-  try {
-    const match = content.match(JSON_REG);
-    return JSON.parse(match[1]);
-  } catch (error) {
-    throw new Error('Unable to derive JSON object in response.');
-  }
 }
 
 export function transformResponse(options) {
@@ -35,8 +31,26 @@ export function transformResponse(options) {
   } else if (output === 'messages') {
     return [...messages, message];
   } else if (output === 'json') {
-    return parse(content);
+    return parseJson(content);
+  } else if (output === 'code') {
+    return parseCode(content);
   } else {
-    throw new Error(`Unknown output type "${output}".`);
+    throw new Error(`No output type provided.`);
+  }
+}
+
+function parseJson(content) {
+  try {
+    return JSON.parse(content.match(JSON_REG)[0]);
+  } catch (error) {
+    throw new Error(`Unable to derive JSON from response:\n\n${content}`);
+  }
+}
+
+function parseCode(content) {
+  try {
+    return content.match(CODE_REG)[1].trim();
+  } catch (error) {
+    throw new Error(`Unable to derive code from response:\n\n${content}`);
   }
 }